+++ src/c45/Pair.java
@@ -1,15 +1,15 @@
 package c45;
 
 public class Pair {
-	int id_attr;
+	double id_attr;
 	String id_class;
 	
-	Pair(int id_attr,String id_class){
+	Pair(double id_attr,String id_class){
 		this.id_attr = id_attr;
 		this.id_class = id_class;	
 	}
 	
-	public int getId_attr() {
+	public double getId_attr() {
 		return id_attr;
 	}
